refactor: comment 관련 컴포넌트를 comments 폴더로 이동
