+++ src/components/comments/CommentEditForm.jsx
@@ -1,7 +1,7 @@
-import { formatDateKRW } from "../utils/formatPrice";
+import { formatDateKRW } from "../../utils/formatPrice";
 import styles from "./CommentEditForm.module.css";
 import { useState } from "react";
-import ProfileCard from "./layout/profileCard/ProfileCard";
+import ProfileCard from "../layout/profileCard/ProfileCard";
 
 const INQUIRE_PLACEHOLDER =
   "개인정보를 공유 및 요청하거나, 명예 훼손, 무단 광고, 불법 정보 유포시 모니터링 후 삭제될 수 있으며, 이에 대한 민형사상 책임은 게시자에게 있습니다.";
