+++ src/components/comments/CommentView.jsx
@@ -1,7 +1,7 @@
-import { formatDateKRW } from "../utils/formatPrice";
+import { formatDateKRW } from "../../utils/formatPrice";
 import styles from "./CommentView.module.css";
-import KebabMenu from "./common/KebabMenu/KebabMenu";
-import ProfileCard from "./layout/profileCard/ProfileCard";
+import KebabMenu from "../common/KebabMenu/KebabMenu";
+import ProfileCard from "../layout/profileCard/ProfileCard";
 
 const CommentView = ({ comment, onEdit, onDelete }) => {
   const dropDownItems = [
